A patch to make the testing framework work in Google
