+++ src/python/grpcio_tests/tests_aio/unit/test_base.py
@@ -12,6 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import os
+import sys
+import subprocess
+
 import asyncio
 import unittest
 import socket
@@ -35,11 +39,50 @@ def _get_free_loopback_tcp_port():
     return tcp_socket, "%s:%s" % (host_target, address_tuple[1])
 
 
+class _Server:
+    """_Server is an wrapper for a sync-server subprocess.
+
+    The synchronous server is executed in another process which initializes
+    implicitly the grpc using the synchronous configuration. Both worlds
+    can not coexist within the same process.
+    """
+
+    def __init__(self, host_and_port):  # pylint: disable=W0621
+        self._host_and_port = host_and_port
+        self._handle = None
+
+    def start(self):
+        assert self._handle is None
+
+        try:
+            from google3.pyglib import resources
+            executable = resources.GetResourceFilename(
+                "google3/third_party/py/grpc/sync_server")
+            args = [executable, '--host_and_port', self._host_and_port]
+        except ImportError:
+            executable = sys.executable
+            directory, _ = os.path.split(os.path.abspath(__file__))
+            filename = directory + '/sync_server.py'
+            args = [
+                executable, filename, '--host_and_port', self._host_and_port
+            ]
+
+        self._handle = subprocess.Popen(args)
+
+    def terminate(self):
+        if not self._handle:
+            return
+
+        self._handle.terminate()
+        self._handle.wait()
+        self._handle = None
+
+
 class AioTestBase(unittest.TestCase):
 
     def setUp(self):
         self._socket, self._target = _get_free_loopback_tcp_port()
-        self._server = sync_server.Server(self._target)
+        self._server = _Server(self._target)
         self._server.start()
         self._loop = asyncio.new_event_loop()
         asyncio.set_event_loop(self._loop)
